backend/schedules-ai/requirements.txt
# Core API
fastapi[standard]>=0.104.1
uvicorn[standard]>=0.24.0
# flask>=2.0.0
cryptography>=44.0.3
hypercorn>=0.14.0
asgiref>=3.4.0

# Database
asyncpg>=0.27.0

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0.1
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utilities
pytz>=2023.3.post1

# LLM Engine Dependencies
requests>=2.31.0
aiohttp>=3.9.1
json5>=0.9.14
torch>=2.4.0

# Constraint Solver
ortools>=9.8

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1

# Linting
ruff>=0.1.9

# JWT Handling
PyJWT>=2.8.0
